requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Task queue
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
boto3==1.34.0

# Selenium and browser automation
selenium==4.15.2
webdriver-manager==4.0.1
selenium-stealth==1.0.6
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
selenium-screenshot==3.0.0

# AI/ML
openai==1.51.2
google-generativeai==0.8.3
python-jobspy==0.0.1
pandas==2.1.4

# Cloud storage and Firestore
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 